added frontend UI

+++ backend/controllers/goalController.js
@@ -1,13 +1,15 @@
 const asyncHandler = require('express-async-handler')
 
+const Goal = require('../models/goalModel')
+const User = require('../models/userModel')
+
 // Desc     Get goals
 // Method   GET
 // route    /api/goals
 // access   Private
 const getGoals = asyncHandler(async (req, res) => {
-  res.status(200).json({
-    message: 'Get goals'
-  })
+  const goals = await Goal.find({ user: req.user.id })
+  res.status(200).json(goals)
 }
 )
 // Desc     Set,Create goals
@@ -19,11 +21,12 @@ const setGoal = asyncHandler(async (req, res) => {
     res.status(400)
     throw new Error('Please add a goal')
   }
-  res.status(200).json(
-    {
-      method: 'post',
-      message: 'Create goals'
-    })
+
+  const goal = await Goal.create({
+    user: req.user.id,
+    text: req.body.text
+  })
+  res.status(200).json(goal)
 }
 )
 // Desc     Update goal
@@ -31,10 +34,22 @@ const setGoal = asyncHandler(async (req, res) => {
 // route    /api/goals/:id
 // access   Private
 const updateGoal = asyncHandler(async (req, res) => {
-  res.status(200).json({
-    method: 'put',
-    message: `Update goals ${req.params.id}`
-  })
+  const goal = await Goal.findById(req.params.id)
+  if (!goal) {
+    res.status(400)
+    throw new Error('Goal not found')
+  }
+  const user = await User.findById(req.user.id)
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+  const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, { new: true })
+  res.status(200).json(updatedGoal)
 }
 )
 // Desc     Delete goal
@@ -42,9 +57,25 @@ const updateGoal = asyncHandler(async (req, res) => {
 // route    /api/goals/:id
 // access   Private
 const deleteGoal = asyncHandler(async (req, res) => {
+  const goal = await Goal.findById(req.params.id)
+
+  if (!goal) {
+    res.status(400)
+    throw new Error('Goal not found')
+  }
+  const user = await User.findById(req.user.id)
+  if (!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+  if (goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+  await Goal.remove(req.params.id)
   res.status(200).json({
     method: 'delete',
-    message: `Delete goals ${req.params.id}`
+    id: req.params.id,
   })
 }
 )
